% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscunidatasets-package.R
\docType{package}
\name{DSCUNIDataSets}
\alias{DSCUNIDataSets}
\alias{DSCUNIDataSets-package}
\title{DSCUNIDataSets: A Comprehensive Collection of University and Engineering Datasets}
\description{
{
This package offers a curated collection of datasets focused on university and engineering research, covering mechanical engineering experiments and clustering studies.Designed for researchers, analysts, and data scientists.
}
}
\details{
DSCUNIDataSets: A Comprehensive Collection of University and Engineering Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of University and Engineering Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/dscclub-uni}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
